3 funcionalidades reparacion

+++ src/main/java/com/eoi/ejemplospringboot/repositories/ReparacionRepository.java
@@ -5,5 +5,6 @@ package com.eoi.ejemplospringboot.repositories;
 import com.eoi.ejemplospringboot.entities.Reparacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReparacionRepository extends JpaRepository<Reparacion, Long> {
+public interface ReparacionRepository extends JpaRepository<Reparacion, Integer> {
+
 }
